# server/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Dragon Math API",
    "version": "1.0.0",
    "description": "The routes the Dragon Academy iOS app uses. Generated from server/contracts/ by `npm run openapi`; do not edit by hand."
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    },
    "schemas": {
...
      "ErrorResponse": {
        "type": "object",
        "properties": {
          "error": {
            "type": "string",
            "description": "Human-readable message, safe to show a kid or parent."
          }
        },
        "required": [
          "error"
        ]
      },
...
      "SpellingListsResponse": {
        "type": "object",
        "properties": {
          "lists": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/SpellingList"
            }
          }
        },
        "required": [
          "lists"
        ]
      },
      "SpellingList": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer"
          },
          "name": {
            "type": "string"
          },
          "child_id": {
            "type": "integer"
          },
          "created_at": {
            "type": "string",
            "nullable": true,
            "description": "ISO 8601 timestamp."
          },
          "updated_at": {
            "type": "string",
            "nullable": true,
            "description": "ISO 8601 timestamp. Changes on every edit."
          },
          "created_by_self": {
            "type": "boolean",
            "description": "True when the caller made the list (false: \"added by a grown-up\")."
          },
          "words": {
            "type": "array",
            "items": {
              "type": "string"
            },
            "description": "Lower-case words, in the order the list was typed."
          },
          "audio_missing": {
            "type": "array",
            "items": {
              "type": "string"
            },
            "description": "Words with no recorded clip; their audio URL 404s."
          },
          "example_sentences": {
            "type": "object",
            "additionalProperties": {
              "type": "string"
            },
            "description": "Word → a sentence using it, for the words that have one."
          }
        },
        "required": [
          "id",
          "name",
          "child_id",
          "created_at",
          "updated_at",
          "created_by_self",
          "words",
          "audio_missing",
          "example_sentences"
        ]
      },
...
    },
...
  },
  "paths": {
...
    "/api/spelling/lists": {
      "get": {
        "operationId": "listSpellingLists",
        "summary": "A child's custom spelling lists, oldest first.",
        "tags": [
          "spelling"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "schema": {
              "type": "integer",
              "minimum": 0,
              "exclusiveMinimum": true,
              "description": "Required for a grown-up (a linked child's id). A child may omit it, or pass their own id."
            },
            "required": false,
            "description": "Required for a grown-up (a linked child's id). A child may omit it, or pass their own id.",
            "name": "child_id",
            "in": "query"
          }
        ],
        "responses": {
          "200": {
            "description": "The lists.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SpellingListsResponse"
                }
              }
            }
          },
          "400": {
            "description": "The request was invalid.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "401": {
            "description": "Missing, malformed or expired credentials.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "403": {
            "description": "Signed in, but not allowed to do this.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/spelling/audio/{word}": {
      "get": {
        "operationId": "getSpellingAudio",
        "summary": "The spoken clip for one spelling word. Public, and cacheable for a week.",
        "tags": [
          "spelling"
        ],
        "parameters": [
          {
            "schema": {
              "type": "string",
              "pattern": "^[A-Za-z]{1,24}(\\.[Mm][Pp]3)?$",
              "description": "The word, optionally with a .mp3 suffix (e.g. dragon.mp3). Case-insensitive."
            },
            "required": true,
            "description": "The word, optionally with a .mp3 suffix (e.g. dragon.mp3). Case-insensitive.",
            "name": "word",
            "in": "path"
          }
        ],
        "responses": {
          "200": {
            "description": "The MP3 clip.",
            "content": {
              "audio/mpeg": {
                "schema": {
                  "type": "string",
                  "format": "binary"
                }
              }
            }
          },
          "400": {
            "description": "The request was invalid.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "Not found.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
...
  }
}